Refactor control plane screens
