+++ src/web/features/control-plane/components/workspaces-screen/WorkspaceMeta.tsx
@@ -1,0 +1,10 @@
+import type { ReactNode } from 'react';
+
+export function WorkspaceMeta({ label, children }: { label: string; children: ReactNode }) {
+  return (
+    <div className="space-y-1">
+      <dt className="text-xs font-semibold uppercase tracking-[0.08em] text-muted-foreground">{label}</dt>
+      <dd className="break-all text-sm text-foreground">{children}</dd>
+    </div>
+  );
+}
