+++ src/web/features/control-plane/components/workspaces-screen/types.ts
@@ -1,0 +1,7 @@
+import type { ControlPlaneState } from '../../../../lib/api';
+
+export type WorkspaceCreateInput = { name: string; anchorRoot: string; setActive: boolean };
+
+export type WorkspaceListItem = ControlPlaneState['workspaces'][number] & {
+  relation: 'attached' | 'known';
+};
